requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
marshmallow==3.20.1
python-dotenv==1.0.0
Werkzeug==3.0.1
python-dateutil==2.8.2
email-validator==2.1.0
bcrypt==4.1.2
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
factory-boy==3.3.0 